+++ tests/devices_test.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+
+python -m pytest -m 'not todo' test_trainer_models.py test_trainer_device.py