+++ tests/http_test.sh
@@ -1,0 +1,4 @@
+#! /bin/bash
+
+python -m pytest -m 'http and not todo' -k 'not daemon_modules_http' && \
+python -m pytest -m 'not todo' test_daemon_modules_http.py